DetectFraudFunction/requirements.txt
azure-functions
azure-cosmos
azure-servicebus
python-dateutil
